Finalized the tweaks to personalization/recommendation integration.  Final edits before public release in T minus 28 minutes.
I want to take a second to say how proud I am of the work that I have put in on this project thus far. Getting to this state is pretty incredible, and I am so excited to take the limitations that my mind had been creating and crush them with the help of the suggestions of my friends. This project is about to get better in ways that I can't even imagine yet!